feat: implement scroll handling with new custom hook
- Added a new hook, useAttachScrollBeyond, to manage scroll behavior and state across components.
- Updated ChatInterface and WelcomeScreen to utilize the new hook for handling scroll events, improving the user experience by ensuring consistent scroll threshold management.
- Refactored AIChatRoot to maintain clean code structure.

These changes enhance the chat interface by providing a centralized way to manage scroll interactions, improving overall usability.
